+++ src/main/java/stock/entity/StockEntity.java
@@ -13,7 +13,7 @@ import static jakarta.persistence.GenerationType.AUTO;
 public class StockEntity {
 	
 	@Id
-	private int stockID;
+	private Long stockID;
 	
 	@Column(unique = true, nullable = false, name = "Stock_Name")
 	private String stockName;
